# By Lucas Casagrande (3) and Nicokuyum (1)
# Via Nicokuyum
* 'master' of https://bitbucket.org/itba/paw-2016b-04:
  Creado otro logback de produccion y agregado al pom padre que ignore el de desarrollo al crear el .war
  Revert "avanzando con hibernate"
  arreglando cambios
  arregle en algunos lados la excepcion de proxy, cambie el map para que se fije si el user del sector es null

Conflicts:
	interfaces/src/main/java/ar/edu/itba/interfaces/BuildingDao.java
	interfaces/src/main/java/ar/edu/itba/interfaces/EmpireDao.java
	interfaces/src/main/java/ar/edu/itba/interfaces/EmpireService.java
	interfaces/src/main/java/ar/edu/itba/interfaces/ResourceDao.java
	interfaces/src/main/java/ar/edu/itba/interfaces/SectorService.java
	model/src/main/java/ar/edu/itba/model/Alert.java
	model/src/main/java/ar/edu/itba/model/Army.java
	model/src/main/java/ar/edu/itba/model/Message.java
	model/src/main/java/ar/edu/itba/model/Resource.java
	model/src/main/java/ar/edu/itba/model/Sector.java
	model/src/main/java/ar/edu/itba/model/TradeOffer.java
	model/src/main/java/ar/edu/itba/model/Troop.java
	model/src/main/java/ar/edu/itba/model/User.java
	persistence/src/main/java/ar/edu/itba/persistence/ArmyHibernateDao.java
	persistence/src/main/java/ar/edu/itba/persistence/ArmyJdbcDao.java
	persistence/src/main/java/ar/edu/itba/persistence/BuildingHibernateDao.java
	persistence/src/main/java/ar/edu/itba/persistence/BuildingJdbcDao.java
	persistence/src/main/java/ar/edu/itba/persistence/CommerceHibernateDao.java
	persistence/src/main/java/ar/edu/itba/persistence/CommerceJdbcDao.java
	persistence/src/main/java/ar/edu/itba/persistence/EmpireHibernateDao.java
	persistence/src/main/java/ar/edu/itba/persistence/EmpireJdbcDao.java
	persistence/src/main/java/ar/edu/itba/persistence/ResourceHibernateDao.java
	persistence/src/main/java/ar/edu/itba/persistence/ResourceJdbcDao.java
	persistence/src/main/java/ar/edu/itba/persistence/TroopJdbcDao.java
	persistence/src/main/java/ar/edu/itba/persistence/UserJdbcDao.java
	persistence/src/test/java/ar/edu/itba/persistence/TerrainJdbcDaoTest.java
	service/src/main/java/ar/edu/itba/service/EmpireServiceImpl.java
	service/src/main/java/ar/edu/itba/service/SectorServiceImpl.java
	webapp/src/main/java/ar/edu/itba/paw/webapp/controller/ArmyController.java
	webapp/src/main/java/ar/edu/itba/paw/webapp/controller/BuildingController.java
	webapp/src/main/java/ar/edu/itba/paw/webapp/controller/HomePageController.java

+++ interfaces/src/main/java/ar/edu/itba/interfaces/UserService.java
@@ -6,11 +6,11 @@ public interface UserService {
 
 	public User findByUsername(String username);
 	
-	public User findById(int id);
+	public User findById(long id);
 	
 	public User create(String username, String password, String email);
 	
-	public String getUsername(int id);
+	public String getUsername(long id);
 
 	public boolean exists(String username);
